+++ src/nksip_subscription.erl
@@ -25,7 +25,7 @@
 -author('Carlos Gonzalez <carlosj.gf@gmail.com>').
 
 -export([field/3, field/2, fields/3, id/1, id/2, dialog_id/1, subscription_id/2]).
--export([get_subscription/2, get_all/0, get_all/2, remote_id/2]).
+-export([get_subscription/2, get_all/0, get_all/2, notify_status/1, remote_id/2]).
 -export_type([id/0, status/0, terminated_reason/0]).
 
 -include("nksip.hrl").
@@ -225,6 +225,60 @@ get_all(AppId, CallId) ->
         nksip_dialog:field(AppId, DlgId, subscriptions)
         || {_, DlgId} <- nksip_call_router:get_all_dialogs(AppId, CallId)
     ]).
+
+%% @private
+-spec notify_status(nksip:request()) ->
+    {active|pending, non_neg_integer()} | 
+    {terminated, nksip_subscription:terminated_reason()}.
+
+notify_status(#sipmsg{}=SipMsg) ->
+    case nksip_sipmsg:header(SipMsg, <<"Subscription-State">>, tokens) of
+        [{Status, Opts}] ->
+            case nksip_lib:get_list(<<"expires">>, Opts) of
+                "" -> 
+                    Expires = undefined;
+                Expires0 ->
+                    case catch list_to_integer(Expires0) of
+                        Expires when is_integer(Expires) -> Expires;
+                        _ -> Expires = undefined
+                    end
+            end,
+            case Status of
+                <<"active">> -> 
+                    {active, Expires};
+                <<"pending">> -> 
+                    {pending, Expires};
+                <<"terminated">> ->
+                    Retry = case nksip_lib:get_value(<<"retry_after">>, Opts) of
+                        undefined ->
+                            undefined;
+                        Retry0 ->
+                            case nksip_lib:to_integer(Retry0) of
+                                Retry1 when is_integer(Retry1), Retry1>=0 -> Retry1;
+                                _ -> undefined
+                            end
+                    end, 
+                    case nksip_lib:get_value(<<"reason">>, Opts) of
+                        undefined -> 
+                            {terminated, undefined};
+                        Reason0 ->
+                            Reason1 = case catch 
+                                binary_to_existing_atom(Reason0, latin1) 
+                            of
+                                {'EXIT', _} -> undefined;
+                                probation -> {probation, Retry};
+                                giveup -> {giveup, Retry};
+                                Reason -> Reason
+                            end,
+                            {terminated, Reason1}
+                    end;
+                _ ->
+                    {terminated, undefined}
+            end;
+        _ ->
+            {terminated, undefined}
+    end.
+